chore(feature/vue-query): 테스트용 2 파일들 삭제
