--- components/mypage/Contents.vue
@@ -1,3 +1,0 @@
-<template></template>
-<script setup></script>
-<style></style>
